adding a identity server to app

+++ src/AuctionService/Controllers/AuctionController.cs
@@ -9,6 +9,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
@@ -54,13 +55,14 @@ namespace AuctionService.Controllers
             return _mapper.Map<AuctionDto>(auction);
 
         }
+        [Authorize]
         [HttpPost]
         public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
         {
             var auction = _mapper.Map<Auction>(auctionDto);
 
             ///todo add current user as seler
-            auction.Seller = "test";
+            auction.Seller = User.Identity.Name;
             _context.Auctions.Add(auction);
 
             var newAuction = _mapper.Map<AuctionDto>(auction);
@@ -75,13 +77,14 @@ namespace AuctionService.Controllers
             return CreatedAtAction(nameof(GetAuctionById),
                  new {auction.Id}, newAuction);
         }
+        [Authorize]
         [HttpPut("{id}")]
         public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
         {
             var auction = await _context.Auctions.Include(x => x.Item).FirstOrDefaultAsync(x =>x.Id ==id);
 
             if(auction == null) return NotFound();
-            //TODO: check seller == username
+            if(auction.Seller != User.Identity.Name) return Forbid();
             auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
             auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
             auction.Item.Color = updateAuctionDto.Color ?? auction.Item.Color;
@@ -94,13 +97,15 @@ namespace AuctionService.Controllers
             if(result) return Ok();
             return BadRequest("Problem saving changes");
         }
+
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteAuction(Guid id)
         {
             var auction = await _context.Auctions.FindAsync(id);
 
             if(auction == null) return NotFound();
-            //Todo check seller == username
+           if(auction.Seller != User.Identity.Name) return Forbid();
 
             _context.Auctions.Remove(auction);
             await _publishEndpoint.Publish<AuctionDeleted>(new {Id = auction.Id.ToString()});
